Refactored package structure.
