+++ piwik-sdk/src/test/java/org/piwik/sdk/testhelper/DefaultTestCase.java
@@ -1,7 +1,9 @@
-package org.piwik.sdk;
+package org.piwik.sdk.testhelper;
 
 import org.junit.Before;
 import org.junit.runner.RunWith;
+import org.piwik.sdk.Piwik;
+import org.piwik.sdk.Tracker;
 import org.robolectric.Robolectric;
 import org.robolectric.annotation.Config;
 
@@ -9,9 +11,9 @@ import java.net.MalformedURLException;
 
 @Config(emulateSdk = 18, manifest = Config.NONE)
 @RunWith(FullEnvTestRunner.class)
-public abstract class PiwikDefaultTest {
+public abstract class DefaultTestCase {
     public Tracker createTracker() throws MalformedURLException {
-        TestPiwikApplication app = (TestPiwikApplication) Robolectric.application;
+        PiwikTestApplication app = (PiwikTestApplication) Robolectric.application;
         return Piwik.getInstance(Robolectric.application).newTracker(app.getTrackerUrl(), app.getSiteId());
     }
