+++ piwik-sdk/src/test/java/org/piwik/sdk/testhelper/FullEnvPackageManager.java
@@ -5,7 +5,7 @@
  * @license https://github.com/piwik/piwik-sdk-android/blob/master/LICENSE BSD-3 Clause
  */
 
-package org.piwik.sdk;
+package org.piwik.sdk.testhelper;
 
 import android.content.ComponentName;
 import android.content.Intent;
