+++ piwik-sdk/src/test/java/org/piwik/sdk/tools/LogyTest.java
@@ -5,7 +5,7 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.piwik.sdk.FullEnvTestRunner;
+import org.piwik.sdk.testhelper.FullEnvTestRunner;
 import org.robolectric.annotation.Config;
